requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
numpy==1.26.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
sklearn==0.0.post11
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
